refactor, add edit-button-skill-card component
