+++ src/presentation/components/organisms/SkillCard.tsx
@@ -1,33 +1,22 @@
-import { Button, Paper, Typography } from "@mui/material";
+import { Paper, Typography } from "@mui/material";
 import { FC } from "react";
-import { useNavigate } from "react-router-dom";
 import { ISkill } from "../../../domain/Skill";
-import { useAppSelector } from "../../hooks/reduxHooks";
 
 type Props = {
   skill: ISkill;
+  children?: React.ReactNode;
 };
 
 const SkillCard: FC<Props> = (props) => {
-  const navigate = useNavigate();
-
-  const { userId: stateUserId } = useAppSelector((state) => state.auth);
-  const { title, userId, _id: skillId } = props.skill;
-
-  const editClickHandler = () => {
-    navigate(`/skills/${skillId}/edit`);
-  };
+  const { skill, children } = props;
+  const { title } = skill;
 
   return (
     <Paper className="px-5 py-3" variant="outlined">
       <Typography variant="body1" sx={{ mb: 2 }}>
         {title}
       </Typography>
-      {stateUserId === userId && (
-        <Button variant="outlined" onClick={editClickHandler}>
-          Edit
-        </Button>
-      )}
+      {children}
     </Paper>
   );
 };
